refactor: Improved code organization and separation of concerns
